requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0
fastapi==0.136.0
filelock @ file:///home/task_176943957172477/croot/filelock_1769440496252/work
fsspec==2026.3.0
gmpy2 @ file:///home/task_176474799776688/conda-bld/gmpy2_1764748047044/work
greenlet==3.4.0
groq==1.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.5.3
lxml==6.1.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///home/task_176537858778399/conda-bld/networkx_1765378626589/work
numpy==1.26.4
packaging==26.0
pgvector==0.4.2
psycopg2-binary==2.9.11
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
PyMuPDF==1.27.2.2
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML @ file:///home/task_176346545528585/conda-bld/pyyaml_1763465495007/work
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
sympy @ file:///home/task_176476213407811/conda-bld/sympy_1764762918097/work
threadpoolctl==3.6.0
tokenizers==0.22.2
torch @ file:///home/justinosoares/Documentos/projects/python/api-context-ai/torch-2.4.0%2Bcpu-cp311-cp311-linux_x86_64.whl#sha256=14a7a8b595347dddca594f9e448b93ce68ce4f871acbd32cf04bda7c03664c0c
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
uvicorn==0.44.0
